Convert defaultRenderer tests to use snapshots so they're testing interface instead of implementation. Convert file name to snake case.
