+++ src/components/value_renderer/default_renderer/default_renderer.js
